Replace Stream by search params
