--- ui/src/app/(dashboard)/stream/@slideOver/default.tsx
@@ -1,3 +1,0 @@
-export default function SlideOverDefault() {
-  return null;
-}
