+++ ui/src/app/(dashboard)/stream/StreamDetails.tsx
